requirements.txt
Django==4.2.*
django-ninja
redis==5.0.1
django-redis==5.4.0
